+++ _legacy/constants/index.ts
@@ -1,0 +1,401 @@
+const TILE_SIZE = 8;
+const SAMPLE_POLYGON: Array<[number, number]> = [
+  [0.65125, 0.73125],
+  [0.6275, 0.7375],
+  [0.61, 0.735],
+  [0.5925, 0.72625],
+  [0.57875, 0.71875],
+  [0.56625, 0.71375],
+  [0.54875, 0.715],
+  [0.53125, 0.72],
+  [0.515, 0.7225],
+  [0.50125, 0.71625],
+  [0.48125, 0.7175],
+  [0.46375, 0.72625],
+  [0.44875, 0.73],
+  [0.4375, 0.7325],
+  [0.42625, 0.7325],
+  [0.425, 0.73625],
+  [0.43375, 0.755],
+  [0.44375, 0.76375],
+  [0.45625, 0.7725],
+  [0.4625, 0.77625],
+  [0.48, 0.7825],
+  [0.49875, 0.785],
+  [0.5125, 0.785],
+  [0.5275, 0.78125],
+  [0.53625, 0.78125],
+  [0.55125, 0.78375],
+  [0.56625, 0.785],
+  [0.5775, 0.785],
+  [0.5875, 0.785],
+  [0.60125, 0.7775],
+  [0.6175, 0.76375],
+  [0.6375, 0.74375],
+  [0.64625, 0.73625],
+  [0.65125, 0.73125],
+  [0.6375, 0.71875],
+  [0.6225, 0.7075],
+  [0.60875, 0.695],
+  [0.6, 0.68625],
+  [0.585, 0.67125],
+  [0.57875, 0.6675],
+  [0.5675, 0.66125],
+  [0.55875, 0.6625],
+  [0.54875, 0.665],
+  [0.535, 0.67375],
+  [0.52625, 0.67375],
+  [0.5125, 0.6675],
+  [0.50625, 0.66375],
+  [0.49125, 0.66625],
+  [0.47375, 0.67375],
+  [0.45875, 0.68875],
+  [0.44375, 0.7],
+  [0.43375, 0.70875],
+  [0.42125, 0.7175],
+  [0.4125, 0.71875],
+  [0.40375, 0.71875],
+  [0.4, 0.71625],
+  [0.405, 0.72125],
+  [0.41, 0.72375],
+  [0.41875, 0.7275],
+  [0.435, 0.72625],
+  [0.45125, 0.7225],
+  [0.47625, 0.6975],
+  [0.49125, 0.685],
+  [0.505, 0.66],
+  [0.5125, 0.63875],
+  [0.51375, 0.61875],
+  [0.50625, 0.60125],
+  [0.4975, 0.595],
+  [0.485, 0.59125],
+  [0.47375, 0.58875],
+  [0.4675, 0.58875],
+  [0.4625, 0.59125],
+  [0.4675, 0.6],
+  [0.4775, 0.60375],
+  [0.49, 0.61],
+  [0.505, 0.61125],
+  [0.5225, 0.6125],
+  [0.53375, 0.6125],
+  [0.545, 0.60625],
+  [0.55625, 0.59625],
+  [0.56625, 0.5925],
+  [0.575, 0.5925],
+  [0.5825, 0.59375],
+  [0.565, 0.605],
+  [0.55375, 0.61],
+  [0.54125, 0.61125],
+  [0.5325, 0.6125],
+  [0.565, 0.6125],
+  [0.58375, 0.61],
+  [0.59375, 0.60125],
+  [0.59875, 0.59375],
+  [0.59875, 0.58375],
+  [0.59625, 0.5725],
+  [0.5925, 0.565],
+  [0.58375, 0.55375],
+  [0.57, 0.54],
+  [0.565, 0.525],
+  [0.56125, 0.49],
+  [0.55875, 0.39],
+  [0.56125, 0.3375],
+  [0.5725, 0.29625],
+  [0.585, 0.2675],
+  [0.59625, 0.24625],
+  [0.615, 0.22375],
+  [0.64375, 0.20375],
+  [0.66875, 0.195],
+  [0.6875, 0.19],
+  [0.70875, 0.18875],
+  [0.7325, 0.19],
+  [0.7525, 0.1925],
+  [0.78875, 0.20375],
+  [0.81875, 0.21875],
+  [0.835, 0.2325],
+  [0.84375, 0.24625],
+  [0.85, 0.2675],
+  [0.8525, 0.2875],
+  [0.84875, 0.30375],
+  [0.83875, 0.32375],
+  [0.82625, 0.3375],
+  [0.8025, 0.35375],
+  [0.77125, 0.37375],
+  [0.74375, 0.38],
+  [0.71125, 0.3825],
+  [0.6825, 0.385],
+  [0.6625, 0.38375],
+  [0.64625, 0.38125],
+  [0.62875, 0.37875],
+  [0.615, 0.3725],
+  [0.6075, 0.3675],
+  [0.62625, 0.34375],
+  [0.64125, 0.33125],
+  [0.66125, 0.31375],
+  [0.68875, 0.3],
+  [0.715, 0.295],
+  [0.7525, 0.2925],
+  [0.77375, 0.29375],
+  [0.79625, 0.3],
+  [0.8125, 0.31],
+  [0.8275, 0.32375],
+  [0.83875, 0.33625],
+  [0.85, 0.35375],
+  [0.86875, 0.39125],
+  [0.87875, 0.4475],
+  [0.87625, 0.49375],
+  [0.87625, 0.54375],
+  [0.865, 0.5975],
+  [0.84625, 0.6525],
+  [0.82875, 0.69625],
+  [0.81125, 0.7375],
+  [0.77375, 0.7925],
+  [0.75, 0.815],
+  [0.70625, 0.85375],
+  [0.67125, 0.87875],
+  [0.6375, 0.89875],
+  [0.6075, 0.91125],
+  [0.57625, 0.9225],
+  [0.54375, 0.925],
+  [0.49875, 0.9275],
+  [0.48, 0.92375],
+  [0.45, 0.9175],
+  [0.4225, 0.90125],
+  [0.38625, 0.88375],
+  [0.3475, 0.8525],
+  [0.32, 0.8325],
+  [0.305, 0.825],
+  [0.295, 0.82125],
+  [0.28125, 0.81875],
+  [0.26125, 0.81875],
+  [0.245, 0.82375],
+  [0.22, 0.83125],
+  [0.20625, 0.84125],
+  [0.17875, 0.84625],
+  [0.1675, 0.8475],
+  [0.1575, 0.8525],
+  [0.1375, 0.85],
+  [0.12, 0.84375],
+  [0.09625, 0.83375],
+  [0.0675, 0.8175],
+  [0.04625, 0.8075],
+  [0.035, 0.80125],
+  [0.02625, 0.7975],
+  [0.0225, 0.7975],
+  [0.05625, 0.77625],
+  [0.09125, 0.7575],
+  [0.11375, 0.74625],
+  [0.1475, 0.7425],
+  [0.1625, 0.7425],
+  [0.18, 0.745],
+  [0.20125, 0.74875],
+  [0.21875, 0.75625],
+  [0.24, 0.76625],
+  [0.26375, 0.78],
+  [0.2825, 0.79375],
+  [0.30375, 0.81125],
+  [0.32125, 0.82625],
+  [0.32875, 0.83625],
+  [0.27375, 0.76875],
+  [0.2525, 0.72875],
+  [0.25125, 0.70875],
+  [0.265, 0.685],
+  [0.29375, 0.66625],
+  [0.32375, 0.6525],
+  [0.35375, 0.635],
+  [0.375, 0.61875],
+  [0.39625, 0.59375],
+  [0.4175, 0.5625],
+  [0.4275, 0.53625],
+  [0.43125, 0.5175],
+  [0.43125, 0.51125],
+  [0.4125, 0.52875],
+  [0.3975, 0.53375],
+  [0.3675, 0.53625],
+  [0.36, 0.53625],
+  [0.33375, 0.5375],
+  [0.315, 0.5425],
+  [0.29625, 0.5475],
+  [0.27875, 0.5625],
+  [0.25875, 0.58375],
+  [0.245, 0.60125],
+  [0.2375, 0.62125],
+  [0.2375, 0.645],
+  [0.2375, 0.6775],
+  [0.24125, 0.69625],
+  [0.24625, 0.71375],
+  [0.24875, 0.72375],
+  [0.25125, 0.72625],
+  [0.2325, 0.65],
+  [0.23125, 0.605],
+  [0.2325, 0.575],
+  [0.23375, 0.53625],
+  [0.235, 0.51],
+  [0.23625, 0.49625],
+  [0.26875, 0.5],
+  [0.28875, 0.49625],
+  [0.31375, 0.48875],
+  [0.34, 0.47875],
+  [0.36375, 0.465],
+  [0.3775, 0.4475],
+  [0.39875, 0.43],
+  [0.42375, 0.41375],
+  [0.4475, 0.40125],
+  [0.47, 0.38375],
+  [0.485, 0.3625],
+  [0.49375, 0.33625],
+  [0.4975, 0.31],
+  [0.48375, 0.27875],
+  [0.46625, 0.26],
+  [0.45, 0.25125],
+  [0.4325, 0.25],
+  [0.41, 0.25125],
+  [0.39125, 0.25875],
+  [0.35875, 0.27875],
+  [0.335, 0.2875],
+  [0.315, 0.29125],
+  [0.2975, 0.29],
+  [0.28125, 0.28125],
+  [0.2575, 0.265],
+  [0.23, 0.245],
+  [0.20375, 0.23125],
+  [0.1825, 0.225],
+  [0.1675, 0.22375],
+  [0.15875, 0.2225],
+  [0.18625, 0.20375],
+  [0.23375, 0.19],
+  [0.2625, 0.185],
+  [0.29625, 0.1775],
+  [0.3275, 0.1775],
+  [0.35625, 0.18375],
+  [0.375, 0.195],
+  [0.3925, 0.21125],
+  [0.40875, 0.22875],
+  [0.41875, 0.24875],
+  [0.42375, 0.2725],
+  [0.4275, 0.30375],
+  [0.43125, 0.33125],
+  [0.42375, 0.35125],
+  [0.40625, 0.36875],
+  [0.38, 0.37875],
+  [0.3625, 0.38375],
+  [0.335, 0.38],
+  [0.32, 0.365],
+  [0.3025, 0.34625],
+  [0.2875, 0.33875],
+  [0.27375, 0.33625],
+  [0.2375, 0.33625],
+  [0.21625, 0.32875],
+  [0.19625, 0.32125],
+  [0.1775, 0.3075],
+  [0.16125, 0.29375],
+  [0.14875, 0.27375],
+  [0.13875, 0.25625],
+  [0.135, 0.24],
+  [0.1425, 0.235],
+  [0.16125, 0.23625],
+  [0.1725, 0.24375],
+  [0.18875, 0.25875],
+  [0.1975, 0.2725],
+  [0.21375, 0.28875],
+  [0.22375, 0.2975],
+  [0.24125, 0.30375],
+  [0.255, 0.3075],
+  [0.2775, 0.3075],
+  [0.29875, 0.3075],
+  [0.30875, 0.31375],
+  [0.32125, 0.325],
+  [0.3275, 0.34625],
+  [0.3275, 0.3725],
+  [0.325, 0.395],
+  [0.32125, 0.4075],
+  [0.31625, 0.43125],
+  [0.31375, 0.45],
+  [0.305, 0.47],
+  [0.29, 0.485],
+  [0.27375, 0.4875],
+  [0.24875, 0.4875],
+  [0.2325, 0.485],
+  [0.21875, 0.475],
+  [0.21, 0.4625],
+  [0.1975, 0.44],
+  [0.18875, 0.405],
+  [0.1825, 0.37625],
+  [0.17375, 0.35125],
+  [0.16, 0.32375],
+  [0.14375, 0.2975],
+  [0.125, 0.27625],
+  [0.115, 0.26375],
+  [0.0975, 0.25375],
+  [0.08625, 0.25125],
+  [0.07375, 0.2525],
+  [0.06625, 0.255],
+  [0.0625, 0.26875],
+  [0.0625, 0.2825],
+  [0.07125, 0.295],
+  [0.0775, 0.3025],
+  [0.0925, 0.3175],
+  [0.10125, 0.3275],
+  [0.1125, 0.3375],
+  [0.1225, 0.35375],
+  [0.12625, 0.3725],
+  [0.12875, 0.3925],
+  [0.13375, 0.41125],
+  [0.14125, 0.425],
+  [0.1525, 0.44],
+  [0.1675, 0.455],
+  [0.18, 0.46625],
+  [0.1975, 0.47625],
+  [0.215, 0.4875],
+  [0.2275, 0.48875],
+  [0.24375, 0.49375],
+  [0.25375, 0.495],
+  [0.27875, 0.49125],
+  [0.2925, 0.4875],
+  [0.31125, 0.47375],
+  [0.34, 0.44375],
+  [0.34375, 0.415],
+  [0.35125, 0.38875],
+  [0.35625, 0.35625],
+  [0.36375, 0.32375],
+  [0.36875, 0.2925],
+  [0.3675, 0.26],
+  [0.3625, 0.24125],
+  [0.355, 0.22625],
+  [0.34125, 0.21125],
+  [0.33, 0.20375],
+  [0.30875, 0.19625],
+  [0.29125, 0.195],
+  [0.27625, 0.1975],
+  [0.265, 0.2],
+  [0.2575, 0.20875],
+  [0.2575, 0.22375],
+  [0.26, 0.23875],
+  [0.26875, 0.24875],
+  [0.27875, 0.25375],
+  [0.2925, 0.25875],
+  [0.30625, 0.26125],
+  [0.32125, 0.25875],
+  [0.32875, 0.255],
+  [0.3375, 0.24875],
+  [0.3425, 0.24125],
+  [0.34125, 0.23125],
+  [0.32875, 0.22],
+  [0.31875, 0.215],
+  [0.3025, 0.21],
+  [0.2925, 0.20875],
+  [0.2825, 0.2125],
+  [0.28125, 0.225],
+  [0.28875, 0.23375],
+  [0.2975, 0.23625],
+  [0.30875, 0.2375],
+  [0.315, 0.23375],
+  [0.315, 0.2275],
+  [0.31125, 0.2225],
+  [0.305, 0.22],
+  [0.29875, 0.22125],
+  [0.29625, 0.225],
+  [0.29625, 0.22625],
+];
+
+export { SAMPLE_POLYGON, TILE_SIZE };
