+++ _legacy/utils/CaveGenerator.ts
