+++ _legacy/utils/Lighting.ts
