+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.0.0",
   "private": false,
   "license": "MIT",
-  "homepage": "https://simdaesoo.github.io/Ellie/build",
   "keywords": [
     "game",
     "platform",
@@ -25,29 +24,28 @@
     "url": "https://github.com/SimDaeSoo/Ellie.git"
   },
   "description": "2D Game Engine Using Typescript (Javascript) the goal of this project is to create efficient engine for huge tile platform 2D game like terraria.",
-  "dependencies": {
-    "@pixi/layers": "^1.0.8",
+  "devDependencies": {
     "@types/node": "^12.0.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
     "@types/react-router-dom": "^5.3.0",
     "@types/seedrandom": "^3.0.1",
     "@types/stats.js": "^0.17.0",
-    "pixi-filters": "^4.1.4",
-    "pixi.js": "^6.1.3",
+    "react-app-rewired": "^2.1.8",
+    "react-scripts": "4.0.3",
+    "typescript": "^4.1.2"
+  },
+  "dependencies": {
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
     "rsuite": "^4.10.3",
     "seedrandom": "^3.0.5",
-    "stats.js": "^0.17.0",
-    "typescript": "^4.1.2"
+    "stats.js": "^0.17.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "eject": "react-scripts eject"
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build"
   },
   "eslintConfig": {
     "extends": [
