+++ public/WebWorker.js
@@ -1,0 +1,6 @@
+onmessage = (e) => {
+  const { data } = e;
+  console.log(data);
+}
+
+postMessage('init');
