+++ public/index.html
@@ -6,7 +6,8 @@
   <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <meta name="theme-color" content="#000000" />
-  <meta name="description" content="2D Game Engine Using Typescript (Javascript) the goal of this project is to create efficient engine for huge tile platform 2D game like terraria." />
+  <meta name="description"
+    content="2D Game Engine Using Typescript (Javascript) the goal of this project is to create efficient engine for huge tile platform 2D game like terraria." />
   <link rel="apple-touch-icon" href="%PUBLIC_URL%/icon_192x192.png" />
   <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
   <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/index.css">
