--- src/pages/LineIntersection.tsx
@@ -1,71 +1,0 @@
-import * as PIXI from 'pixi.js';
-import { useEffect } from 'react';
-import { Line } from '../interfaces';
-import { getClientSize, setRenderer } from '../utils';
-import { drawLines, drawPoints } from '../utils/Graphics';
-import { createLabel } from '../utils/Label';
-import * as Collision from '../utils/Collision';
-
-const LineIntersection = ({
-  setCallback,
-}: {
-  setCallback: (callback: (x: number, y: number) => void) => void;
-}) => {
-  useEffect(() => {
-    // Set Click Callback
-    setCallback((_x: number, _y: number) => {});
-
-    // Render
-    const [clientWidth, clientHeight] = getClientSize();
-    const { stage } = setRenderer();
-    const graphics = new PIXI.Graphics();
-    const [quarterWidth, quarterHeight] = [
-      Math.floor(clientWidth / 4),
-      Math.floor(clientHeight / 4),
-    ];
-
-    // Lines
-    const lineA: Line = [
-      [quarterWidth, quarterHeight],
-      [quarterWidth * 3, quarterHeight * 3],
-    ];
-    const lineB: Line = [
-      [quarterWidth * 3, quarterHeight],
-      [quarterWidth, quarterHeight * 3],
-    ];
-
-    // Draw Points
-    const points = [...lineA, ...lineB];
-    stage.addChild(graphics);
-
-    for (const point of points) {
-      const { container: pointContainer } = createLabel(`(${point})`);
-      pointContainer.position.set(...point);
-      stage.addChild(pointContainer);
-    }
-
-    graphics.clear();
-    drawLines(graphics, [lineA, lineB], { width: 1, color: 0x009900 });
-    drawPoints(graphics, points, 2);
-
-    // Line Intersection
-    const collisionDt: number = Collision.lineIntersection(lineA, lineB);
-    const collisionPoint: [number, number] = [
-      lineA[0][0] + (lineA[1][0] - lineA[0][0]) * collisionDt,
-      lineA[0][1] + (lineA[1][1] - lineA[0][1]) * collisionDt,
-    ];
-    drawPoints(graphics, [collisionPoint], 4, 0xffaa66);
-
-    const { container: pointContainer } = createLabel(
-      `Intersect at (${collisionPoint[0].toFixed(
-        2
-      )},${collisionPoint[1].toFixed(2)})`
-    );
-    pointContainer.position.set(...collisionPoint);
-    stage.addChild(pointContainer);
-  }, [setCallback]);
-
-  return <></>;
-};
-
-export default LineIntersection;
